axora/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axora"
version = "2.0.0"
description = "AXORA — Autonomous AI Developer Agent CLI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "AXORA" }]

dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "cryptography>=42.0",
    "psutil>=5.9",
    "pydantic>=2.0",
    "anthropic>=0.28",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "black", "ruff"]

[project.scripts]
axora = "axora.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["axora*"]
